[#5] feat - opening creation modal
